update offscreen rendering
