+++ scene/stage.js
@@ -47,6 +47,7 @@ class Stage {
 			depth: true,
 			antialias: true,
 			alpha: false,
+			powerPreference: 'high-performance',
 		});
